+++ lib/key_convert.h
@@ -4,6 +4,6 @@
 #include "stdint.h"
 
 uint8_t convert_key(uint8_t key_code);
-uint8_t set_zero(uint8_t input, uint8_t grade);
+uint8_t set_zero(uint8_t input, uint16_t grade);
 
 #endif
